backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
motor==3.6.0
pydantic==2.9.2
pydantic-settings==2.6.0
bcrypt==4.2.0
pyjwt==2.9.0
python-multipart==0.0.12
email-validator==2.2.0
